Basic auto-differentiation engine that works with scalar tensors and the four arithmetic operations + negation.

+++ src/tensor/iterator/util.rs
@@ -1,10 +1,10 @@
-use crate::traits::haslength::HasLength;
+use crate::util::haslength::HasLength;
 
 // returns a pair of vectors: one indexed by indices and the other with the remaining elements
 pub(super) fn split_by_indices<T, I>(data: &[T], indices: I) -> (Vec<T>, Vec<T>)
 where
     T: Copy,
-    I: HasLength + IntoIterator<Item=isize>,
+    I: HasLength + IntoIterator<Item=usize>,
 {
     let max_index = data.len();
     let mut bitset = vec![false; max_index];
@@ -13,11 +13,11 @@ where
     let mut remaining = Vec::with_capacity(max_index - indices.len());
 
     for index in indices {
-        if bitset[index as usize] {
+        if bitset[index] {
             panic!("duplicate index encountered");
         }
         
-        bitset[index as usize] = true;
+        bitset[index] = true;
     }
 
     for (&value, &bit) in data.iter().zip(bitset.iter()) {
